--- feature/history/src/main/java/com/potato/history/model/HistoryGroupedItem.kt
@@ -1,8 +1,0 @@
-package com.potato.history.model
-
-import com.poptato.domain.model.response.history.HistoryItemModel
-
-data class HistoryGroupedItem(
-    val date: String,
-    val items: List<HistoryItemModel>
-)
