--- feature/history/src/main/java/com/potato/history/model/HistoryItemModel.kt
@@ -1,8 +1,0 @@
-package com.potato.history.model
-
-data class HistoryItemModel(
-    val id: Long,
-    val title: String,
-    val date: String,
-    val isChecked: Boolean
-)
